OSIS-5403 : Refactor (not working)
