+++ base/management/commands/load_sql_scripts.py
@@ -4,7 +4,6 @@ import sys
 from typing import Dict, List
 
 import attr
-from django.core.management.base import BaseCommand
 from django.db import connection
 
 LOCK_TEMPLATE = """
@@ -15,11 +14,13 @@ LOCK_TEMPLATE = """
 """
 
 
+@attr.s()
 class SQLLockStatement:
-    lock_template = attr.ib(type=str, default=None)
+    lock_template = attr.ib(type=str, default=LOCK_TEMPLATE)
     lock_mode = attr.ib(type=str, default=None)
 
 
+@attr.s()
 class LoadSQLFilesToExecute:
     subfolder = attr.ib(type=str, default=None)
     scripts_path = attr.ib(type=str, default=None)
@@ -46,15 +47,13 @@ class LoadSQLFilesToExecute:
         return script_strings
 
 
-class LoadSqlCommand(BaseCommand, SQLLockStatement, LoadSQLFilesToExecute):
+class LoadSQL(SQLLockStatement, LoadSQLFilesToExecute):
     backoffice: str = 'backoffice/'
     cursor = None
     tablename_regex = None
 
     def __init__(self):
         super().__init__()
-        self.scripts_path = self.backoffice + self.subfolder
-        self.lock_template = LOCK_TEMPLATE
         self.cursor = connection.cursor()
 
     def _get_table_name(self, sql_string: str):
@@ -66,3 +65,32 @@ class LoadSqlCommand(BaseCommand, SQLLockStatement, LoadSQLFilesToExecute):
         except Exception:
             sys.exit("#### Error: Unable to get table name from sql file ####")
 
+
+class LoadSQLTriggers(LoadSQL):
+
+    def __init__(self):
+        self.subfolder = 'triggers/'
+        self.scripts_path = self.backoffice + self.subfolder
+        self.tablename_regex = r'CREATE TRIGGER[\S\s]*(public..*)'
+        self.lock_mode = 'SHARE ROW EXCLUSIVE'
+        super().__init__()
+
+    def load_triggers(self):
+        trigger_strings = self.load_scripts()
+        print("## Loading triggers from {} ##".format(self.scripts_path))
+        for trigger in trigger_strings:
+            self.load_trigger(trigger)
+        print("## Loading triggers finished ##")
+
+    def load_trigger(self, trigger: Dict[str, str]):
+        table_name = self._get_table_name(trigger['script_string'])
+        print("YOLOOO", self.lock_mode)
+        sql_script = self.lock_template.format(
+            table_to_lock=table_name,
+            sql_script=trigger['script_string'],
+            lock_mode=self.lock_mode
+        )
+        print("# Load trigger from {filename} #".format(filename=trigger['filename']))
+        print("# Table {tablename} LOCKED #".format(tablename=table_name))
+        self.cursor.execute(sql_script)
+        print("# Table {tablename} UNLOCKED #".format(tablename=table_name))
